feat: Organizing the schemas into common and specific ones
